refactor: extract global idle manager from chat to site-wide
- Move createIdleManager to app/src/scripts/idle-manager.ts
- Create global instance in BaseLayout.ts with document-level activity
  detection (mousemove, keydown, touchstart on entire page)
- Dispatch site:idle / site:active events instead of chat:idle / chat:active
- Chat now listens for site events instead of owning idle detection
- Any page interaction resets idle, not just chat panel activity
