+++ app/src/scripts/idle-manager.ts
